src2/server/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.8.0
ffmpeg==1.4
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fsspec==2023.10.0
gTTS==2.4.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
openai==1.1.1
PyAudio==0.2.14
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.0
sympy==1.12
torch==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
